pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lte_pm_platform"
version = "0.1.0"
description = "Governed ZTE LTE PM ingestion and coverage analysis platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "psycopg[binary]>=3.2,<4.0",
  "python-dotenv>=1.0,<2.0",
  "typer>=0.16,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<0.12",
]

[project.scripts]
lte-pm = "lte_pm_platform.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
]
